Code cleanup and rearrangement
