+++ src/minswap/models/__init__.py
@@ -11,9 +11,17 @@ from minswap.models.common import (  # noqa: F401
     AssetHistoryReference,
     AssetIdentity,
     Assets,
+    OrderDatum,
     Output,
+    PlutusFullAddress,
+    PlutusNone,
+    PoolDatum,
     PoolTransactionReference,
+    SwapExactIn,
+    SwapExactOut,
     Transaction,
     TxContentUtxo,
     TxIn,
+    ZapIn,
+    asset_to_value,
 )
